Fix finch package name.
