+++ samples/server/petstore/finch/src/main/scala/Server.scala
@@ -1,22 +1,23 @@
-package io.swagger.petstore
+package io.swagger
 
 import io.finch._
 import io.finch.circe._
-import io.circe.{ Decoder, ObjectEncoder }
+import io.circe.{Decoder, ObjectEncoder}
 import io.circe.generic.auto._
 import io.circe.generic.semiauto
 import io.circe.generic.semiauto._
 import io.circe.java8.time._
 import com.twitter.finagle.Http
 import com.twitter.finagle.util.LoadService
-import com.twitter.util.{ Await, Future }
+import com.twitter.util.{Await, Future}
+
 
 class Server {
 
-  // Loads implementation defined in resources/META-INF/services/io.swagger.petstore.DataAccessor
+  // Loads implementation defined in resources/META-INF/services/io.swagger.DataAccessor
   val db = LoadService[DataAccessor]() match {
     case accessor :: _ => accessor
-    case _ => new DataAccessor {}
+    case _ => new DataAccessor { }
   }
 
   val service = endpoint.makeService(db)
