+++ samples/server/petstore/finch/src/main/scala/io/swagger/models/User.scala
@@ -1,0 +1,36 @@
+package io.swagger.models
+
+import io.circe._
+import io.finch.circe._
+import io.circe.generic.semiauto._
+import io.circe.java8.time._
+import io.swagger._
+
+/**
+ * A User who is purchasing from the pet store
+ * @param id 
+ * @param username 
+ * @param firstName 
+ * @param lastName 
+ * @param email 
+ * @param password 
+ * @param phone 
+ * @param userStatus User Status
+ */
+case class User(id: Option[Long],
+                username: Option[String],
+                firstName: Option[String],
+                lastName: Option[String],
+                email: Option[String],
+                password: Option[String],
+                phone: Option[String],
+                userStatus: Option[Int]
+                )
+
+object User {
+    /**
+     * Creates the codec for converting User from and to JSON.
+     */
+    implicit val decoder: Decoder[User] = deriveDecoder
+    implicit val encoder: ObjectEncoder[User] = deriveEncoder
+}
