--- samples/server/petstore/finch/src/main/scala/io/swagger/petstore/models/Tag.scala
@@ -1,25 +1,0 @@
-package io.swagger.petstore.models
-
-import io.circe._
-import io.finch.circe._
-import io.circe.generic.semiauto._
-import io.circe.java8.time._
-import io.swagger.petstore._
-
-/**
- * A tag for a pet
- * @param id
- * @param name
- */
-case class Tag(
-  id: Option[Long],
-  name: Option[String]
-)
-
-object Tag {
-  /**
-   * Creates the codec for converting Tag from and to JSON.
-   */
-  implicit val decoder: Decoder[Tag] = deriveDecoder
-  implicit val encoder: ObjectEncoder[Tag] = deriveEncoder
-}
