--- samples/server/petstore/finch/src/main/scala/io/swagger/petstore/models/User.scala
@@ -1,37 +1,0 @@
-package io.swagger.petstore.models
-
-import io.circe._
-import io.finch.circe._
-import io.circe.generic.semiauto._
-import io.circe.java8.time._
-import io.swagger.petstore._
-
-/**
- * A User who is purchasing from the pet store
- * @param id
- * @param username
- * @param firstName
- * @param lastName
- * @param email
- * @param password
- * @param phone
- * @param userStatus User Status
- */
-case class User(
-  id: Option[Long],
-  username: Option[String],
-  firstName: Option[String],
-  lastName: Option[String],
-  email: Option[String],
-  password: Option[String],
-  phone: Option[String],
-  userStatus: Option[Int]
-)
-
-object User {
-  /**
-   * Creates the codec for converting User from and to JSON.
-   */
-  implicit val decoder: Decoder[User] = deriveDecoder
-  implicit val encoder: ObjectEncoder[User] = deriveEncoder
-}
